+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project
 adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Ability to select from available swapchain surface formats when creating an `EventLoop`
+
 ## [0.9.0] - 2023-09-07
 
 ### Fixed
@@ -354,7 +360,7 @@ _See [#25](https://github.com/attackgoat/screen-13/pull/25) for migration detail
   platforms and require no bare-metal graphics API knowledge
 - "Hello, world!" example using a bitmapped font
 
-[Unreleased]: https://github.com/attackgoat/screen-13/compare/5acb1257fcdc4505a2a4e264c5e64a0558bf68aa...HEAD
+[Unreleased]: https://github.com/attackgoat/screen-13/compare/3997cc65b87cbcedbe380e3b23ce4a37f7c7c320...HEAD
 [0.1.0]: https://crates.io/crates/screen-13/0.1.0
 [0.2.0]: https://crates.io/crates/screen-13/0.2.0
 [0.3.0]: https://crates.io/crates/screen-13/0.3.0